Create display.mjs
